Update widget width argument to accept an array.
